requirements.txt
langchain
langchain-community
langchain-openai
langgraph
faiss-cpu
dashscope
python-dotenv
flake8
fastapi
uvicorn
pydantic
tavily-python
langchain-pinecone
pinecone-client
pypdf